enable PostTrackingAction for gamma nuclear (#473)
This PR fixes the handling of gamma nuclear reactions:

Before, the steps that undergo gamma-nuclear reactions were not marked
as the final steps, which means that the PostTrackingAction was
erroneously not called. Now, they are marked as final steps, which then
activates the calling of the UserPostTrackingAction. Note: As the
returned Steps are handled before the leaked tracks (and the
gamma-nuclear reaction will be invoked as a leaked _track_) the step
information of the last step does not yet reflect the interaction.
Therefore, when filling the G4 step information, the kinetic energy is
set correctly to 0, as the interaction would later do.

For the split kernels, the returning of the step of the gamma-nuclear
reaction was missing entirely.

This PR also unified the `surviveFlag` and the `isLastStep`, which were
not both needed. It is replaced by a clearer `trackSurvives`.

It was verified that this PR
- [ ] Changes physics results
- [x] Does not change physics results
Note: it does not change the physics results within the CI and examples,
because they use only energy deposition directly. This PR would fix the
kinetic energy for gamma nuclear reactions, thus if an SD code would
rely on using kinetic energy, it would affect the physics

+++ src/AdePTGeant4Integration.cpp
@@ -491,7 +491,11 @@ void AdePTGeant4Integration::ProcessGPUStep(GPUHit const &hit, bool const callUs
 
   // If this was the last step of a track, the hostTrackData of that track can be safely deleted.
   // Note: This deletes the AdePT-owned UserTrackInfo data
-  if (hit.fLastStepOfTrack) {
+  // Exception: gamma-nuclear steps (fParticleType == 2 and fStepLimProcessId == 3):
+  // Steps are processed before the leaked tracks, and the hostTrackData is still used for invoking
+  // gamma-nuclear when the track is returned to the host (although it will die then). Thus, the hostTrackData must be
+  // kept alive until the invokation of the gamma-nuclear reaction
+  if (hit.fLastStepOfTrack && !(hit.fParticleType == 2 && hit.fStepLimProcessId == 3)) {
     fHostTrackDataMapper->removeTrack(hit.fTrackID);
   }
 }
@@ -797,6 +801,16 @@ void AdePTGeant4Integration::FillG4Step(GPUHit const *aGPUHit, G4Step *aG4Step,
   aPostStepPoint->SetCharge(aTrack->GetParticleDefinition()->GetPDGCharge()); // Real data
   // aPostStepPoint->SetMagneticMoment(0);                                                            // Missing data
   // aPostStepPoint->SetWeight(0);                                                                    // Missing data
+
+  // adjust for gamma-nuclear steps:
+  // As the steps are processed before the leaked tracks, the step has not undergone the gamma-nuclear reaction yet.
+  // Therefore, the kinetic energy for the track and postStepPoint must be set by hand to 0
+  if (aGPUHit->fParticleType == 2 && aGPUHit->fStepLimProcessId == 3) {
+    aTrack->SetKineticEnergy(0.);
+    // aTrack->SetVelocity(0.);              // Not set in other cases, so also not set here
+    aPostStepPoint->SetKineticEnergy(0.);
+    // aPostStepPoint->SetVelocity(0.);      // Not set in other cases, so also not set here
+  }
 
   // Last, call tracking and stepping actions
   // call UserSteppingAction if required
